Replacing ProfileHandler in download_and_profile_util.py
