+++ server/opendp_apps/dataverses/download_and_profile_util.py
@@ -15,15 +15,17 @@ import json
 
 from django.core.serializers.json import DjangoJSONEncoder
 
+from opendp_apps.analysis.models import DepositorSetupInfo
 from opendp_apps.async_messages.websocket_message import WebsocketMessage
 from opendp_apps.async_messages import static_vals as async_static
 
 from opendp_apps.model_helpers.basic_err_check import BasicErrCheck
-from opendp_apps.dataset.models import DataSetInfo #, DataverseFileInfo
+from opendp_apps.dataset.models import DataSetInfo
 
 from opendp_apps.dataverses.dataverse_download_handler import DataverseDownloadHandler
-from opendp_apps.profiler.profile_handler import ProfileHandler
-from opendp_apps.profiler import static_vals as pstatic
+from opendp_apps.profiler.csv_reader import CsvReader
+from opendp_apps.profiler.dataset_info_updater import DataSetInfoUpdater
+from opendp_apps.profiler.profile_runner import run_profile
 
 
 class DownloadAndProfileUtil(BasicErrCheck):
@@ -144,29 +146,28 @@ class DownloadAndProfileUtil(BasicErrCheck):
         self.send_websocket_success_msg('The Dataverse file has been copied.')
         return True
 
-
     def profile_file(self):
         """"Profile the file"""
-        if self.has_error():
+        try:
+            ds_info = DataSetInfo.objects.get(object_id=self.dataset_object_id)
+            filefield = ds_info.source_file
+        except DataSetInfo.DoesNotExist:
+            filefield = None
+        try:
+            df = CsvReader(filefield.path).read()
+        except Exception as ex:
+            self.add_err_msg(str(ex))
+            self.send_websocket_profiler_err_msg(str(ex))
+            if self.dataset_object_id:
+                dataset_info = DataSetInfo.objects.get(object_id=self.dataset_object_id)
+                dataset_info_updater = DataSetInfoUpdater(dataset_info)
+                dataset_info_updater.update_step(DepositorSetupInfo.DepositorSteps.STEP_9300_PROFILING_FAILED)
             return
-
-        params = {pstatic.KEY_DATASET_IS_DJANGO_FILEFIELD: True,
-                  pstatic.KEY_DATASET_OBJECT_ID: self.dataset_info.object_id}
-
-        ph = ProfileHandler(dataset_pointer=self.dataset_info.source_file,
-                            **params)
-
-        if ph.has_error():
-            user_msg = ph.get_err_msg()
-            self.add_err_msg(user_msg)
-            self.send_websocket_profiler_err_msg(user_msg)
-            return
-
+        ph = run_profile(df, self.dataset_object_id)
 
         #self.data_profile = ph.get_data_profile()
         self.profile_variables = ph.get_profile_variables()
         profile_str = json.dumps(self.profile_variables, cls=DjangoJSONEncoder, indent=4)
-
 
         self.send_websocket_success_msg('Profile complete!',
                                         profile_str=profile_str)
